allow a list of admins to post images of faces
This is a temp fix for the false negatives that Rekognition falslu identifies as faces. So an admin can repost the image.

+++ TelegramPrivacyBot.py
@@ -12,6 +12,7 @@ from datetime import datetime
 #import MessageEntity
 
 def image(update, context):
+    admin_list = ['jojo786', 'Muaaza'] #List of telegram users that can bypass the rules and still post
     date = datetime.now().strftime("%d/%m/%Y %H:%M:%S")
     print(date + " - Start processing image:")
 
@@ -35,7 +36,7 @@ def image(update, context):
     os.remove("image.jpg") #delete the file once its already processed
     
     found_face = str({len(response['FaceDetails'])})
-    if response['FaceDetails']:
+    if response['FaceDetails'] and (chat_user.username not in admin_list): #if there was a face found, and the person posting is NOT an admin, then delete
         print(date + " - Found " + found_face + " faces in image from user " + str(chat_user.username) + " in group " + str(chat_group) + ", going to delete")
         #context.bot.send_message(chat_id=chat_id, text="Found " + found_face + " faces, deleting...")
         context.bot.delete_message(chat_id=chat_id, message_id=update.message.message_id)
